Moved wout(), etc. into utility/world.hpp
